requirements.txt
duffel-api
hydra-core
openai